Improve infection logic
Breaks backwards compatibility since remove_client and update_client won't work on older clients.

Adds launchers and loaders:
The only goal of launchers is to run the stager code, this small piece of code will connect to the server which will then create an encrypted payload using the client's key. Once the payload is created the server will wrap the payload in a loader who decides how the payload is loaded on the client system, this allows for multiple ways of persistence. Lastly the loader is sent back to the stager who will run it.

The way payloads are encrypted using the client's key is based on what Patrick Wardle said in one of his talks, I probably implemented it wrong though.