Update to v4.0.0
This commit is unstable, there's still a lot of things to re-add.
From now on this we'll be focusing on controlling many machines at once.
Commands are now hidden inside 404 pages (from BlackHat's "Hiding In Plain Sight")
Encryption support is dropped for now.
There's a new and improved module system.